tests/wpt/web-platform-tests/workers/constructors/SharedWorker/interface-objects.html

<!doctype html>
<title>expected interface objects/constructors</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
(async_test()).step(function() {
  var expected = 'XMLHttpRequest WebSocket EventSource MessageChannel Worker SharedWorker ApplicationCache'.split(' ');
  var supported = [];
  for (var i = 0; i < expected.length; ++i) {
  if (expected[i] in window)
     supported.push(expected[i]);
  }
  var worker = new SharedWorker('interface-objects.js');
  worker.port.start();
  worker.port.postMessage(supported);
  worker.port.addEventListener('message', this.step_func(function(e) {
    assert_equals(e.data, '');
    this.done();
  }), false);
});
</script>
